use mmcv.init_dist (#1851)
