+++ mmdet/apis/__init__.py
@@ -1,10 +1,9 @@
-from .env import get_root_logger, init_dist, set_random_seed
 from .inference import (async_inference_detector, inference_detector,
                         init_detector, show_result, show_result_pyplot)
-from .train import train_detector
+from .train import get_root_logger, set_random_seed, train_detector
 
 __all__ = [
-    'async_inference_detector', 'init_dist', 'get_root_logger',
-    'set_random_seed', 'train_detector', 'init_detector', 'inference_detector',
-    'show_result', 'show_result_pyplot'
+    'get_root_logger', 'set_random_seed', 'train_detector', 'init_detector',
+    'async_inference_detector', 'inference_detector', 'show_result',
+    'show_result_pyplot'
 ]
