+++ requirements.txt
@@ -1,10 +1,10 @@
-mmcv>=0.2.10
-numpy
+albumentations>=0.3.2
+imagecorruptions
 matplotlib
+mmcv>=0.2.15
+numpy
+pycocotools
 six
 terminaltables
-pycocotools
 torch>=1.1
 torchvision
-imagecorruptions
-albumentations>=0.3.2
